perf(hist): float histogram cells, double reductions (decision 50)
HistCell is {float, float}: per-cell sums are bounded by node size and
gradients arrive as float, so cell storage carries only per-cell rounding.
Reductions that cross cells — totals(), fill_prefix, the node finder's
running left sums — accumulate in double and convert once at the store.
The CUDA engine already proved the shape (float per-chunk, double merge).

Paired on M2 (8 threads): base cell fit 15.3 -> 13.2s, 8M x 100 53.3 ->
45.7s (populate 1.23-1.25x); R2 identical to six decimals at both scales;
eval baseline moved 5.4e-6 (re-pinned). Histogram pool, partial slabs, and
prefix scratch all halve.

The switch flushed out a latent guard gap: f32 sibling-subtraction noise
pushes degenerate cuts (every row one side) past the gain>0 gate; empty-
child covers then made SHAP contributions NaN. demote_empty_splits (and
the leafwise equivalent) now convert any split whose partition leaves a
child empty back to a leaf — row counts are ground truth at any precision.

+++ src/cuda/histogram_engine.cu
@@ -389,8 +389,10 @@ struct CudaHistogramEngine::Impl
                                        .bin_id       = static_cast<bin_id_t>(b.bin),
                                        .default_left = b.dl != 0,
                                        .valid        = true};
-            child_sums[2 * i]       = {.sum_grad = b.gL, .sum_hess = b.hL};
-            child_sums[(2 * i) + 1] = {.sum_grad = b.gR, .sum_hess = b.hR};
+            child_sums[2 * i]       = {.sum_grad = static_cast<float>(b.gL),
+                                       .sum_hess = static_cast<float>(b.hL)};
+            child_sums[(2 * i) + 1] = {.sum_grad = static_cast<float>(b.gR),
+                                       .sum_hess = static_cast<float>(b.hR)};
         }
     }
 
@@ -831,11 +833,13 @@ void CudaHistogramEngine::find_level_split(Dataset const & /*ds*/,
     }
     for (size_t i = 0; i < n; ++i)
     {
-        out[i]                  = split;
-        child_sums[2 * i]       = {.sum_grad = im.level_child.host[(4 * i) + 0],
-                                   .sum_hess = im.level_child.host[(4 * i) + 1]};
-        child_sums[(2 * i) + 1] = {.sum_grad = im.level_child.host[(4 * i) + 2],
-                                   .sum_hess = im.level_child.host[(4 * i) + 3]};
+        out[i]            = split;
+        child_sums[2 * i] = {
+            .sum_grad = static_cast<float>(im.level_child.host[(4 * i) + 0]),
+            .sum_hess = static_cast<float>(im.level_child.host[(4 * i) + 1])};
+        child_sums[(2 * i) + 1] = {
+            .sum_grad = static_cast<float>(im.level_child.host[(4 * i) + 2]),
+            .sum_hess = static_cast<float>(im.level_child.host[(4 * i) + 3])};
     }
     lap(prof.unpack_s);
 }
